Change instancing test scene and fix some warnings

+++ src/mesh.c
@@ -66,7 +66,7 @@ void draw_mesh_with_shader(Mesh const *mesh, Shader const *shader) {
 
     for (usize i = 0; i < mesh->textures_len; ++i) {
         TextureMaterialType const material = mesh->textures[i].material;
-        assert((int) material <= ARRAY_LEN(count));
+        assert((int) material <= (int) ARRAY_LEN(count));
         assert((int) material > 0); // TextureMaterialType_None = 0
 
         // @Todo: store a char const *material_name in the Texture struct, and use string
